backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
sqlalchemy
databases
aiosqlite
tabula-py